feat(Tutorials): add instructions on how to update to unity 2020
Due to Unity 2020 not supporting the old legacy XR system, it means
that the project no longer works on Unity 2020 and needs the newer
Tilia packages for the Unity XR Plugin system for it to work.

This tutorial goes through the steps required to get it working in
Unity 2020. It also shows how to update to the new Unity input system.